First pass at chapter 10, lexer only.
I'm not about to do XML, so I'm doing scheme and tokenizing to a sexp.
Then I'm using racket's xml library to convert from the sexp to XML
automatically. Unfortunately, our text comparison tool is stupid so I
need to pass the output through a normalizer to clean it up.